COM-1043 logging user on reset password
