+++ src/core/pages/signin/Authenticate.vue
@@ -31,12 +31,11 @@
 
 <script>
 import get from 'lodash/get';
-import Users from '@api/Users';
-import Customers from '@api/Customers';
 import CompaniHeader from '@components/CompaniHeader';
 import Input from '@components/form/Input';
 import { NotifyNegative } from '@components/popup/notify';
-import { HELPER, AUXILIARY_ROLES, AUXILIARY_WITHOUT_COMPANY } from '@data/constants';
+import { AUXILIARY_ROLES, AUXILIARY_WITHOUT_COMPANY } from '@data/constants';
+import { loggingMixin } from '@mixins/loggingMixin';
 
 export default {
   metaInfo: {
@@ -51,6 +50,7 @@ export default {
     'compani-header': CompaniHeader,
     'ni-input': Input,
   },
+  mixins: [loggingMixin],
   data () {
     return {
       credentials: {
@@ -76,32 +76,7 @@ export default {
   methods: {
     async submit () {
       try {
-        const authenticationPayload = {
-          email: this.credentials.email.toLowerCase(),
-          password: this.credentials.password,
-        };
-        const auth = await Users.authenticate(authenticationPayload);
-
-        const expiresInDays = parseInt(auth.expiresIn / 3600 / 24, 10) >= 1
-          ? parseInt(auth.expiresIn / 3600 / 24, 10)
-          : 1;
-        const options = { path: '/', expires: expiresInDays, secure: process.env.NODE_ENV === 'production' };
-        this.$q.cookies.set('alenvi_token', auth.token, options);
-        this.$q.cookies.set('alenvi_token_expires_in', auth.expiresIn, options);
-        this.$q.cookies.set('refresh_token', auth.refreshToken, { ...options, expires: 365 });
-        this.$q.cookies.set('user_id', auth.user._id, options);
-        await this.$store.dispatch('main/getLoggedUser', this.$q.cookies.get('user_id'));
-
-        if (this.$route.query.from) return this.$router.replace({ path: this.$route.query.from });
-
-        if (this.userRole === HELPER) {
-          const customer = await Customers.getById(this.loggedUser.customers[0]._id);
-          this.$store.commit('customer/saveCustomer', customer);
-          this.$router.replace({ name: 'customer agenda' });
-        } else if (this.isAuxiliaryWithoutCompany) {
-          this.$router.replace({ name: 'client account info', params: { id: this.loggedUser._id } });
-        } else if (this.isAuxiliary) this.$router.replace({ name: 'auxiliary agenda' });
-        else this.$router.replace({ name: 'auxiliaries directory' });
+        this.loggingUser({ email: this.credentials.email.toLowerCase(), password: this.credentials.password });
       } catch (e) {
         NotifyNegative('Impossible de se connecter.');
         console.error(e);
